API

Equations of motion

TraceParticles.lorentzforce! — Function

lorentzforce!(du, u, p, t)

Equations of motion described by the Lorentz Force in 3 dimensions. The parameters p contains the charge and mass of the particle, and an interpolation functor giving the magnetic and electric field by passing the position coordinates.

TraceParticles.guidingcentreapproximation! — Function

guidingcentreapproximation!(du, u, p, t)

The equations of motion of a charged particle in collisionless plasma under the guiding centre approximation (GCA). Compared to the full motion (described by the Lorentz force), the statevector u is reduced from 6 DoF to 4, namely the guiding centre position and the guiding centre velocity parallel to the magnetic field, [Rx, Ry, Rz, vparal], respectively.

The parameters p are the particle charge, mass and magnetic moment (which is assumed to be constant in the GCA), along with an interpolation functor/functor taking the arguments (x, y, z, t) and giving the magnetic and electric field at an arbitrary position. The function/functor should return two AbstractVectors, representing the 3-component electric and magnetic field, respectievely. If the function/functor returns StaticArrays, fieldgradients will also return StaticArrays.

The function uses ForwardDiff and Interpolations to calculate gradients.

TraceParticles.hybridgcafo! — Function

hybridgcafo!(du, u, p, t)

A hybrid EoM. Runs either the guiding centre approximation or full orbit integration depending on the parameter switch.

Parameter containers

TraceParticles.FullOrbitParams — Type

FullOrbitParams

A container for the parameters used in a particle simulation solving the Lorentz force directly, and the initial state of the particles are drawn using rejection sampling giving them a statistical weight. The struct also contains a terminationcode that indicates what kind of callback terminated the particle trajectory.

TraceParticles.GCAParams — Type

GCAParams

A container for the parameters used in a guiding centre approximation simulation where the initial state of the particles are drawn using rejection sampling and they have a statistical weight. The struct also contains a terminationcode that indicates what kind of callback terminated the particle trajectory.

TraceParticles.HybridParams — Type

HybridParams

Parameter container for a hybrid GCA/FO particle simulation. The eomid determines which EoM to use and the random number generator rng is used to determine the particle phase when switching from GCA to full orbit integration.

TraceParticles.HybridParamsWithDetection — Type

HybridParamsWithDetection

Parameter container for a hybrid GCA/FO particle simulation. It is similar to HybridParams but with switch times and respective EoM-IDs to use for post processing. It also has a maximum number of switches field.

Enumerations

TraceParticles.EoMID — Module

EoMID

FullOrbit = 0 GuidingCentreApproximation = 1

TraceParticles.TerminationCode — Module

TerminationCode

NotTerminated = 0 OutOfBounds = 1 MagneticGradient = 2 MagneticCurvature = 3 ParallelElectricField = 4 Relativistic = 5 MaxSwitches = 6

TraceParticles.HybridScheme — Module

HybridScheme

Default = 0 SaveSwitches = 1

Whether a hybrid GCA/full-orbit run records each switch between the two sets of equations. SaveSwitches selects HybridParamsWithDetection over HybridParams; it is chosen by save_switchinfo in the parameters.

Problem functions

The prob_func of an EnsembleProblem – how each individual particle problem is created.

TraceParticles.PredefinedICs — Type

struct PredefinedICs
    u0
    tspan
    params

Methods

(::PredefinedICs)(prob, ctx)

Remakes the problem with predefined initial conditions u0[ctx.sim_id], time span tspan[ctx.sim_id], and parameters params[ctx.sim_id].

TraceParticles.SampleFullOrbit — Type

SampleFullOrbit(sampler, tf)

An EnsembleProblem problem function for full-orbit particles whose initial state vector is drawn from sampler, integrated until tf.

TraceParticles.SampleGCA — Type

SampleGCA(sampler, tf)

An EnsembleProblem problem function for guiding-centre particles whose initial state vector is drawn from sampler, integrated until tf.

TraceParticles.SampleHybrid — Type

SampleHybrid(sampler, tf, initialeomid, hybridscheme)

An EnsembleProblem problem function for hybrid GCA/full-orbit particles whose initial state vector is drawn from sampler, integrated until tf.

initialeomid selects which equations of motion a particle starts on, and hybridscheme whether the switches are recorded (see HybridParamsWithDetection) or not.

Output functions

The output_func of an EnsembleProblem – what is kept from each problem solution.

TraceParticles.output_func_lightweight_gca — Function

output_func_lightweight_gca(sol, _)

Output function for ensemble simulations using EnsembleProblem from SciML with the guidingcentreapproximation! equation.

Returns only the initial and final state of the particle, in addition to its charge, mass, magnetic moment, statistical weight, sampling-rejections, SciML-returncode, and TraceParticles.TerminationCode.

TraceParticles.output_func_lightweight_fo — Function

output_func_lightweight_fo(sol, _)

Output function for ensemble simulations using EnsembleProblem from SciML with the lorentzforce! equation.

Returns only the initial and final state of the particle, in addition to its charge, mass, magnetic moment, statistical weight, sampling-rejections, SciML-returncode, and TraceParticles.TerminationCode

TraceParticles.output_func_lightweight_hybrid — Function

output_func_lightweight_hybrid(sol, _)

Output function for ensemble simulations using EnsembleProblem from SciML with the hybridgcafo! equation.

Returns only the initial and final state of the particle, in addition to its charge, mass, magnetic moment, statistical weight, sampling-rejections, SciML-returncode, TraceParticles.TerminationCode, initial and final EoMID, and the number of switches.

TraceParticles.output_func_max_lightweight — Function

output_func_max_lightweight(sol, ctx)

Output function for ensemble simulations using EnsembleProblem from SciML.

Finds the maximum larmor radius, minimum field length and maximum scales ratio of the particle's trajectory using the inherent interpolation function in the particle solution. Also returns the initial and final state of the particle, in addition to its charge, mass and magnetic moment.

We set the required "rerun" return argument to false.

Reduction functions

The reduction of an EnsembleProblem – what happens to a batch of solved problems.

TraceParticles.SaveBatchAsHDF5 — Type

SaveBatchAsHDF5(expdir, expname, batchsize, nbatches)

Reduction functor that saves the batch as an HDF5 file. The struct instance is callable and its fields are

expdir::String: The directory where the data is saved.

expname::String: The name of the experiment.

suffix::String: A suffix to the filename.

batchnr::Int: The current batch number.

batchsize::Int: The size of the batch.

nbatches::Int: The total number of batches.

metadata::Dict{<:String, <:Any}: A dictionary of user-specified metadata.

TraceParticles.get_filename — Function

get_filename(reduction::SaveBatchAsHDF5)

Returns the filename for the current batch from the reduction functor.

TraceParticles.batch_statistics — Function

batch_statistics(batch)

Returns some statistics about the current batch.

Callbacks

Conditions and affects to use in callbacks, for handling events during the particle trajectories. Conditions decide when something happens; affects decide what.

Conditions

TraceParticles.OutOfBoundsCondition — Type

struct OutOfBoundsCondition{
    T1<:NTuple{N,Tuple{Real,Real}} where {N},
    T2<:AbstractVector,
}

Fields

bounds::T1

u_idxs::T2

Returns true if the particle's statevector variable u_idxs[i] is outside the bounds given by bounds[i].

E.g. if the particle has the statevector (x, y, z, vx, vy, vz), then setting bounds = ( (0.0, 0.1), (1.0, 2.0) ) and u_idxs=[1,3] will make the condition return true if the particle's x is outside (0.0, 1.0) or if the particle's z is outside (1.0, 2.0).

TraceParticles.MagneticGradientCondition — Type

MagneticGradientCondition

Callback condition for checking GCA assumption. Returns true if the ratio between the particle Larmor radius and the characteristic length of the magnetic field is higher than a tolerance.

TraceParticles.MagneticCurvatureCondition — Type

MagneticCurvatureCondition

Fields

tolerance

A callback condition that returns true if the ratio between the particle's Larmor radius and curvature radius of the local magnetic field is larger than the tolerance.

TraceParticles.ParallelElectricFieldCondition — Type

ParallelElectricFieldCondition

Fields

tolerance

A callback condition that returns true if the local electric field component parallel to the magnetic field is larger than the electric field component perpendicular to the magnetic field by a factor of tolerance.

TraceParticles.RelativisticCondition — Type

RelativisticCondition

Callback condition for checking if the particle is relativistic. Returns true if the particle's kinetic energy is a user defined fraction of the rest energy.

TraceParticles.RelativisticConditionGCA — Type

RelativisticConditionGCA

Callback condition for checking if the GCA particle is relativistic. Returns true if the particle's kinetic energy is a user defined fraction of the rest energy. Perpendicular drifts other than E cross B drift are neglected.

TraceParticles.RelativisticConditionHybrid — Type

RelativisticConditionHybrid

Callback condition for checking if the particle is relativistic. Returns true if the particle's kinetic energy is a user defined fraction of the rest energy. Perpendicular drifts other than E cross B drift are neglected.

TraceParticles.HybridSwitchCondition — Type

HybridSwitchCondition

Callback condition for checking whether it's time to switch equations of motion when using the hybrid GCA/FO method. The condition returns true if the current equations are full orbit and the GCA is determined to be valid, or if the current equations are GCA but the GCA is determined to be invalid.

The GCA is determined to be invalid in the following cases:

The ratio between the particle's Larmor radius and the characteristic length of the magnetic field strength is larger than tol_gradient.

The ratio between the particle's Larmor radius and the curvature radius of the magnetic field is larger than tol_curvature.

The ratio between the parallel and perpendicular electric field components is larger than tol_efield.

The GCA is determined to be valid if the points above are false even with an additional factor switchback_tol applied to the tolerances. The switchback tolerance is used to avoid rapid oscillations between the GCA and full orbit.

Affects

TraceParticles.outofboundsaffect! — Function

outofboundsnaffect!(integrator)

Callback affect which terminates the integration and sets the termination message to TerminationCode.OutOfBounds.

TraceParticles.magneticgradientaffect! — Function

magneticgradientaffect!(integrator)

Callback affect which terminates the integration and sets the return message to TerminationCode.MagneticGradient.

TraceParticles.magneticcurvatureaffect! — Function

magneticcurvatureaffect!(integrator)

Callback affect which terminates the integration and sets the return message to TerminationCode.MagneticCurvature.

TraceParticles.parallelelectricfieldaffect! — Function

parallelelectricfieldaffect!(integrator)

Callback affect which terminates the integration and sets the return message to TerminationCode.ParallelElectricField.

TraceParticles.relativisticaffect! — Function

relativisticaffect!(integrator)

Callback affect which terminates the integration and sets the return message to TerminationCode.Relativistic.

TraceParticles.hybridswitchaffect! — Function

hybridswitchaffect!(integrator)

Callback affect that checks the switch parameter and switches the EoM, either from full orbit integration to the guiding centre approximation or the opposite.

If the switch is from full orbit to the guiding centre approximation, the parallel velocity and the magnetic moment is comptuted from the particle state vector and the local magnetic field.

If the switch is from the guiding centre approximation to full orbit solution, the particle is given a gyration phase angle according to integrator.p.getphase(integrator) (probably random), and a velocity vector computed from the magnetic moment, parallel velocity, local magnetic field, and the given gyration phase.

See also get_guidingcentre and get_fullorbit.

TraceParticles.hybridswitchaffect_withdetection! — Function

hybridswitchaffect_withdetection!(integrator)

Callback affect that checks the switch parameter and switches the EoM, either from full orbit integration to the guiding centre approximation or the opposite. Also records the time of the switch and the new EoM ID in the switchtimes and neweomid arrays, respectively.

Physics

TraceParticles.get_guidingcentre — Function

get_guidingcentre(
    pos          ::AbstractVector,
    vel          ::AbstractVector,
    magneticfield::AbstractVector,
    electricfield::AbstractVector,
    charge       ::Real,
    mass         ::Real
    )

Calculates and returns the guiding-centre position, parallel velocity, and magnetic moment of a charged particle in a electromagnetic field.

TraceParticles.get_fullorbit — Function

get_fullorbit(
    magneticfield::AbstractVector,
    electricfield::AbstractVector,
    R            ::AbstractVector,
    vparal       ::Real,
    μ            ::Real,
    mass         ::Real,
    charge       ::Real,
    phaseangle   ::Real,
    )

Get the position and velocity from the guiding centre position R, parallel velocity vparal, and magnetic moment μ of a charged particle with mass and charge in an electromagnetic field.

Requires an arbitrary phaseangle of the gyromotion because we are going from 5 parameters to 6. The phase angle is with respect to vector perpendicular to the magnetic field and guiding centre position vector.

TraceParticles.scalesratio — Function

Calculates the ratio between the Larmor radius and the characteristic field length of the magnetic field.

Methods

scalesratio(B, ∇B, vperp, q, m)
# Using interpolation object
scalesratio(R, mass, charge, magneticmoment, emfields)
scalesratio(pos, vel, mass, charge, emfields)

Arguments

R, the guiding centre location.

pos, the 3 component particle position.

vel, the 3 component particle velocity.

∇B, the magnitude of the gradient of the magnetic field

B, the magnetic field strength.

vperp, the particle velocity component perpendicular to the magnetic field.

q, the particle charge.

m, the particle mass.

emfields, a functor that returns a tuple of the electric and magnetic field vectors at a given position, in the form

TraceParticles.magneticcurvatureratio — Function

magneticcurvatureratio(R, t, charge, mass, magneticmoment, emfields)
magneticcurvatureratio(pos, vel, t, charge, mass, emfields)

Calculate the ratio between the particle's Larmor radius and the radius of curvature of the local magnetic field.

Statistics

TraceParticles.mhdsample — Function

mhdsample(
    mass::Real,
    rng::AbstractRNG,
    proposal_distr,
    target_distr,
    tg_itp,
    xmin::Real, xmax::Real,
    ymin::Real, ymax::Real,
    zmin::Real, zmax::Real,
    tmin::Real, tmax::Real,
    max_value::Real,
)

Draw x, y, z positions and time t from a proposal_distribution using rejection sampling. Draws the 3D velocity vector from a Maxwellian distribution with a temperature given by tg_itp(x,y,z). Evaluates the statistical importance weight of the particle using the target_distribution.

TraceParticles.rejectionsample — Function

rejectionsample(
    rng::AbstractRNG,
    precision::DataType,
    target::Any,
    maxvalue::Real,
    xmin::Real, xmax::Real,
    ymin::Real, ymax::Real,
    zmin::Real, zmax::Real,
    tmin::Real, tmax::Real,
)

Sample a point from the 4D target-distribution using rejection sampling. The target function must be callable with the signature (x, y, z, t). The maxvalue determines the upper bound of the uniform "proposal" distribution. Make sure sure that this value is equal to or larger than the maximum possible value of the target distribution. If not, the point will not be a true sample of the target distribution.

The function returns the sampled point and the number of rejections.

The rejection sample algorithm draws independent and identically distributed random variables, at the cost of a potentially large number of rejections. The number of rejections icreases with the number of dimensions.

TraceParticles.maxwellianvelocitysample — Function

maxwellianvelocitysample(
    rng::AbstractRNG,
    temperature_itp::Any,
    mass::Real,
    args...
    ;
    precision::DataType=Float64,
)

Draw a velocity component sample from a Maxwellian distribution with temperature given by a function call with aribtrary arguments, and by a mass.

TraceParticles.MHDSampler — Type

MHDSampler(
    target_distr
    proposal_distr
    tg_itp
    x0 xf y0 yf z0 zf
    tmin tmax
    zbounds
    t0bounds
    max_value
    precision
)

Construct a functor for doing repeated mhdsamples from the same MHD environment.

Methods

(::MHDSampler)(mass, rng)

Return an mhdsample; (x, y, z, vx, vy, vz, t, weight, nrejections).

Post-processing and I/O

TraceParticles.get_observable — Function

Get the observable sym from an ODESolution or a vector of ODESolutions at a single time or over a range of times.

Methods

get_observable(vectorofsolutions, symbol; t, kwargs...)
get_observable(solution, symbol, t; kwargs...)
get_observable(solution, symbol; tspan, times, nsteps, kwargs...)

TraceParticles.h5_getall — Function

h5_getall(filename)
h5_getall(filename, group::String)
h5_getall(filename, batchnr::Int)
h5_getall(filename, batches::AbstractVector)

Return all datasets in the HDF5-file filename as a DataFrame.

TraceParticles.h5_getbatch — Function

h5_getbatch(filename, batchnr)

Returns batch number batchnr as a DataFrame.

TraceParticles.h5_getdataset — Function

h5_getdataset(filename, dataset)
h5_getdataset(filename, dataset, h5group::String)
h5_getdataset(filename, dataset, batchnr::Int)
h5_getdataset(filename, dataset, batches::Vector{<:Int}
h5_getdataset(filename, dataset, batches::Vector{String}

Returns dataset dataset from all batches, the HDF5-group h5group, batchnr or range of batches.

TraceParticles.h5_getenergies — Function

h5_getenergies(filename, args...; units="eV")

Returns the initial and final energies of a set of batches in a HDF5-file. If batches is not specified, the data from all batches are returned.

TraceParticles.h5_getbatchnames — Function

h5_getbatchnames(filename)

Returns the HDF5 group names of the reduction batches.

TraceParticles.save_energy — Function

save_energy(filename, fields_itp)

Calculates the kinetic energy of the initial and final states of an ensamble of test particles.

Particle reruns

TraceParticles.rerun — Function

rerun(idxs, datafile::String, params::TraceParticlesParameters;
      paramsbackup=true, otherparamchanges...)

Re-integrate a chosen set of particles in the electromagnetic field of an existing experiment, keeping their full trajectories instead of the lightweight endpoint summary the ensemble run stores.

The particles are the ones at index idxs in datafile, the HDF5-file the experiment wrote, and params are the parameters it was run with. Any of those parameters can be changed for the rerun by passing it as a keyword argument.

The particle data in datafile is assumed to callable with be an HDF4 file with groups that contains the datasets:

(x0, y0, z0, vparal0, magneticmoment) when running the EoM guidingcentreapproximation!,

(x0, y0, z0, vx0, vy0, vz0) when running the EoM lorentzforce!,

(x0, y0, z0, vx0, vy0, vz0, magneticmoment) when running the EoM hybridgcafo!.

This is the tool for asking how a particle gained its energy: the returned solutions can be fed to get_observable to time-integrate the guiding-centre work channels (:parallelpower, :fermi, :betatron, :polarisationacc), which endpoint data cannot answer.

Example

sol = rerun(1:100, "run.h5", runparams; alg=Vern9())

The high-level interface

TraceParticles.TraceParticlesParameters — Type

TraceParticlesParameters(; kwargs...)

Parameters needed to assemble a TraceParticlesProblem.

All fields are keyword arguments.

Optional callbacks are specified with a tuple of TraceParticles.CallbackSpecs. E.g. see OutOfBounds, Relativistic or HighMagneticGradient.

Providing single-precision input consistently gives a single-precision run.

Mandatory fields

eom: equations of motion,

npart: number of particles,

tf: end time of the integration,

mass, charge: particle mass and charge,

alg: the OrdinaryDiffEq algorithm to integrate with,

maxiters, reltol, abstol: solver controls,

expname: experiment name; also the output subdirectory and file stem,

emfield_file: JLD2 file holding a 6-element vector of callables in the order bx, by, bz, ex, ey, ez,

prob_func: how each particle gets its initial state. Either a specification such as SampleICsFromMHD or ICsFromFile, or any function that EnsembleProblem can call as (prob, ctx).

Optional fields (not a complete list)

output_func: The output the integrator is to save from each particle trajectory. If given it is directly passed as a keyword argument to the SciMLBase.EnsembleProblem. The default is to use output_func_lightweight or output_func_lightweight_hybrid depending on the equations of motion.

reduction: The data reduction of each batch of particles. The default is SaveBatchAsHDF5, which requires the batch to be transformable into a dataframe via DataFrame(batch).

solver_options: extra keyword arguments forwarded verbatim to the SciMLBase.solve call of the ensemble, e.g. (; save_idxs = [1, 2, 3, 4], save_everystep = false). Empty by default. They are splatted after abstol, reltol, maxiters and seed, so repeating one of those here overrides the dedicated field.

callback_specs: tuple of callback specifications, empty by default,

batchsize: number of particles per reduction batch, defaults to npart,

datadir: where the experiment directory is created,

seed: random seed for the ensemble.

paramsbackup::Bool=true: Whether to create a backup of the experiment parameters.

Example

params = TraceParticlesParameters(
    eom = guidingcentreapproximation!,
    npart = 1000,
    tf = 1.0,
    mass = TraceParticles.m_e,
    charge = -TraceParticles.e,
    alg = Tsit5(),
    maxiters = 10_000_000,
    reltol = 3e-8,
    abstol = 1e0,
    expname = "myrun",
    emfield_file = "BE.jld2",
    prob_func = SampleICsFromMHD(
        tg_file = "tg.jld2",
        target_distr_file = "r.jld2",
        xbounds = (1.58e7, 1.63e7),
        ybounds = (1e6, 1e6),
        zbounds = (-6.6e6, -6.1e6),
    ),
    callback_specs = (
        OutOfBounds(x = (1.58e7, 1.63e7), z = (-6.6e6, -6.1e6)),
        Relativistic(fraction = 0.02),
    ),
)

See also TraceParticlesProblem.

TraceParticles.TraceParticlesProblem — Type

TraceParticlesProblem(p::TraceParticlesParameters; logger)

An assembled test-particle simulation, ready to be handed to solve.

The constructor is the assembly step: it loads the electromagnetic field from p.emfield_file, wraps every field component so that it takes (x, y, z, t), initialises the prob_func, picks the output_func and reduction, and creates callbacks implied by p. The result is a SciML EnsembleProblem plus the pieces solve needs alongside it.

Fields

ensemble_prob: the EnsembleProblem to integrate.

callbackset: the CallbackSet built from the termination and equation-of- motion switching parameters in p.

trajectories: number of particles, i.e. p.npart.

batch_size: trajectories per batch, capped at trajectories.

logger: the logger the simulation writes to. Defaults to one that tees to stderr and to a log file in the experiment directory.

See also TraceParticlesParameters.

CommonSolve.solve — Function

solve(prob::TraceParticlesProblem, params::TraceParticlesParameters)

Solve a TraceParticlesProblem given a TraceParticlesParameters instance.

solve(prob::TraceParticlesProblem; kwargs...)

Solve a TraceParticlesProblem and compute the initial and final energy of the particles using save_energy.

prob_func-specifications

prob_func-specifications resolves into one of the problem functions above via init_probfunc.

TraceParticles.SampleICsFromMHD — Type

SampleICsFromMHD(;
    tg_file::String,
    target_distr_file::String,
    xbounds, ybounds, zbounds,
    t0bounds = (0.0, 0.0),
    proposal_distr_file::String = "",
)

Sample initial conditions from an MHD environment.

Positions and time are drawn from a proposal distr. by rejection sampling within the given bounds. The particles are given a statistical weigth that represents the ratio of probailities to be drawn from the target and proposal distribution, weight = Ptarget/Pproposal. The default proposal distribution is the target distribution, giving particles weight=1.

Particle velocities are drawn from a Maxwellian distribution corresponding to the temperature of the drawn position. The temperature is given by the tg_file.

Only filenames of the distributions are stored; the interpolators are loaded by when the problem is assembled in TraceParticleProblem, which is also where the equations of motion decide whether the resulting problem function is a SampleFullOrbit, SampleGCA or SampleHybrid.

TraceParticles.ICsFromFile — Type

ICsFromFile(; ic_file::String)

Take the initial conditions from ic_file, an HDF5 file. Resolves into a PredefinedICs.

Arguments

-ic_file is assumed to be a JLD2 file containing the initial positions, initial velocities, initial times, charge, mass, magnetic moment, statistical weight, and number of rejections (from sampling the condition) for the particles to simulate.

TraceParticles.init_probfunc — Function

init_probfunc(
    probfunc_specifier,
    p::TraceParticlesParameters,
    itp_wrapper,
    fields_itp
)

Resolve probfunc_specifier into a prob_func that EnsembleProblem can call as (prob, ctx).

Anything that is not a recognised specification is returned unchanged, so a user-supplied problem function needs no method here. See SampleICsFromMHD and ICsFromFile examples of specifications that are recognised.

Callback specifiers

TraceParticles.CallbackSpec — Type

CallbackSpec

Supertype of the structs specifying callbacks. See e.g. OutOfBounds, Relativistic and HighMagneticGradient.

TraceParticles.create_callback — Function

create_callback(spec::CallbackSpec, p::TraceParticlesParameters)

Build the callback that enforces spec, matching callback conditions and affects. p supplies the run-wide quantities a criterion may need, such as the particle mass and the equations of motion.

TraceParticles.OutOfBounds — Type

OutOfBounds(; x=nothing, y=nothing, z=nothing, save_positions=(false, true))

Terminate a particle that leaves the given spatial bounds. Each of x, y and z is either nothing (that axis is not checked) or a (lower, upper) tuple. At least one axis must be bounded.

The bounds are only ever compared against the state vector, never multiplied with it, so they are promoted to a common float type among themselves and do not affect the precision the particles are integrated in.

OutOfBounds(x = (1.58e7, 1.63e7), z = (-6.6e6, -6.1e6))

TraceParticles.Relativistic — Type

Relativistic(; fraction=0.12, save_positions=(true, false))

Terminate a particle whose kinetic energy reaches fraction of its rest energy.

TraceParticles.HighMagneticGradient — Type

HighMagneticGradient(; tolerance=1e-4, save_positions=(true, false))

Terminate a particle in a region where the magnetic field varies too rapidly for the guiding-centre approximation, i.e. where the ratio of the Larmor radius to the characteristic field length exceeds tolerance.