Equations of motion

TraceParticles.guidingcentreapproximation! — Function

guidingcentreapproximation!(du, u, p, t)

The equations of motion of a charged particle in collisionless plasma under the guiding centre approximation (GCA). Compared to the full motion (described by the Lorentz force), the statevector u is reduced from 6 DoF to 4, namely the guiding centre position and the guiding centre velocity parallel to the magnetic field, [Rx, Ry, Rz, vparal], respectively.

The parameters p are the particle charge, mass and magnetic moment (which is assumed to be constant in the GCA), along with an interpolation functor/functor taking the arguments (x, y, z, t) and giving the magnetic and electric field at an arbitrary position. The function/functor should return two AbstractVectors, representing the 3-component electric and magnetic field, respectievely. If the function/functor returns StaticArrays, fieldgradients will also return StaticArrays.

The function uses ForwardDiff and Interpolations to calculate gradients.

TraceParticles.lorentzforce! — Function

lorentzforce!(du, u, p, t)

Equations of motion described by the Lorentz Force in 3 dimensions. The parameters p contains the charge and mass of the particle, and an interpolation functor giving the magnetic and electric field by passing the position coordinates.

TraceParticles.hybridgcafo! — Function

hybridgcafo!(du, u, p, t)

A hybrid EoM. Runs either the guiding centre approximation or full orbit integration depending on the parameter switch.

Callbacks

TraceParticles.OutOfBoundsCondition — Type

struct OutOfBoundsCondition{
    T1<:NTuple{N,Tuple{Real,Real}} where {N},
    T2<:AbstractVector,
}

Fields

bounds::T1

u_idxs::T2

Returns true if the particle's statevector variable u_idxs[i] is outside the bounds given by bounds[i].

E.g. if the particle has the statevector (x, y, z, vx, vy, vz), then setting bounds = ( (0.0, 0.1), (1.0, 2.0) ) and u_idxs=[1,3] will make the condition return true if the particle's x is outside (0.0, 1.0) or if the particle's z is outside (1.0, 2.0).

TraceParticles.MagneticGradientCondition — Type

MagneticGradientCondition

Callback condition for checking GCA assumption. Returns true if the ratio between the particle Larmor radius and the characteristic length of the magnetic field is higher than a tolerance switchtol.

TraceParticles.MagneticCurvatureCondition — Type

MagneticCurvatureCondition

Fields

tolerance

A callback condition that returns true if the ratio between the particle's Larmor radius and curvature radius of the local magnetic field is larger than the tolerance.

TraceParticles.ParallelElectricFieldCondition — Type

ParallelElectricFieldCondition

Fields

tolerance

A callback condition that returns true if the local electric field component parallel to the magnetic field is larger than the electric field component perpendicular to the magnetic field by a factor of tolerance.

TraceParticles.RelativisticConditionGCA — Type

RelativisticConditionGCA

Callback condition for checking if the GCA particle is relativistic. Returns true if the particle's kinetic energy is a user defined fraction of the rest energy. Perpendicular drifts other than E cross B drift are neglected.

TraceParticles.outofboundsaffect! — Function

outofboundsnaffect!(integrator)

Callback affect which terminates the integration and sets the termination message to TerminationCode.OutOfBounds.

TraceParticles.magneticgradientaffect! — Function

magneticgradientaffect!(integrator)

Callback affect which terminates the integration and sets the return message to TerminationCode.MagneticGradient.

TraceParticles.magneticcurvatureaffect! — Function

magneticcurvatureaffect!(integrator)

Callback affect which terminates the integration and sets the return message to TerminationCode.MagneticCurvature.

TraceParticles.parallelelectricfieldaffect! — Function

parallelelectricfieldaffect!(integrator)

Callback affect which terminates the integration and sets the return message to TerminationCode.ParallelElectricField.

TraceParticles.relativisticaffect! — Function

relativisticaffect!(integrator)

Callback affect which terminates the integration and sets the return message to TerminationCode.Relativistic.

TraceParticles.HybridSwitchCondition — Type

HybridSwitchCondition

Callback condition for checking whether it's time to switch equations of motion when using the hybrid GCA/FO method. The condition returns true if the current equations are full orbit and the GCA is determined to be valid, or if the current equations are GCA but the GCA is determined to be invalid.

The GCA is determined to be invalid in the following cases:

The ratio between the particle's Larmor radius and the characteristic length of the magnetic field strength is larger than tol_gradient.

The ratio between the particle's Larmor radius and the curvature radius of the magnetic field is larger than tol_curvature.

The ratio between the parallel and perpendicular electric field components is larger than tol_efield.

The GCA is determined to be valid if the points above are false even with an additional factor switchback_tol applied to the tolerances. The switchback tolerance is used to avoid rapid oscillations between the GCA and full orbit.

TraceParticles.hybridswitchaffect! — Function

hybridswitchaffect!(integrator)

Callback affect that checks the switch parameter and switches the EoM, either from full orbit integration to the guiding centre approximation or the opposite.

TraceParticles.TerminationCode — Module

An EnumX module containing termination codes. The codes are set by callback affects that terminate the solver. E.g. the callback affect outofboundsaffect! will terminate the integration and set the parameter terminationcode to TerminationCode.OutOfBounds.

Problem functions

TraceParticles.DposMBvel — Type

Draws x, y, and z positions from a given distribution using rejection sampling.

TraceParticles.PposPvel — Type

Draw initial conditions from an array.

Output functions

TraceParticles.output_func_max_lightweight — Function

output_func_max_lightweight(sol, i)

Output function for ensemble simulations using DifferentialEquations.

Finds the maximum larmor radius, minimum field length and maximum scales ratio of the particle's trajectory using the inherent interpolation function in the particle solution. Also returns the initial and final state of the particle, in addition to its charge, mass and magnetic moment.

We set the required "rerun" return argument to false.

TraceParticles.output_func_lightweight — Function

output_func_lightweight(sol, i)

Output function for ensemble simulations using DifferentialEquations.

Returns only the initial and final state of the particle, in addition to its charge, mass and magnetic moment. We set the required "rerun" return argument to false.

Reduction functions

TraceParticles.SaveBatchAsHDF5 — Type

SaveBatchAsHDF5(expdir, expname, batchsize, nbatches)

Reduction functor that saves the batch as an HDF5 file. The struct instance is callable and its fields are

expdir::String: The directory where the data is saved.

expname::String: The name of the experiment.

suffix::String: A suffix to the filename.

batchnr::Int: The current batch number.

batchsize::Int: The size of the batch.

nbatches::Int: The total number of batches.

metadata::Dict{<:String, <:Any}: A dictionary of user-specified metadata.

TraceParticles.get_filename — Function

get_filename(reduction::SaveBatchAsHDF5)

Returns the filename for the current batch from the reduction functor.

Parameter structs

TraceParticles.GCAParams — Type

GCAParams

A container for the parameters used in a guiding centre approximation simulation where the initial state of the particles are drawn using rejection sampling and they have a statistical weight. The struct also contains a terminationcode that indicates what kind of callback terminated the particle trajectory.

TraceParticles.HybridParams — Type

HybridParams

Parameter container for a hybrid GCA/FO particle simulation. The eomid determines which EoM to use and the random number generator rng is used to determine the particle phase when switching from GCA to full orbit integration.

I/O

TraceParticles.h5_getall — Function

h5_getall(filename)
h5_getall(filename, group::String)
h5_getall(filename, batchnr::Int)
h5_getall(filename, batches::AbstractVector)

Return all datasets in the HDF5-file filename as a DataFrame.

TraceParticles.h5_getbatch — Function

h5_getbatch(filename, batchnr)

Returns batch number batchnr as a DataFrame.

TraceParticles.h5_getdataset — Function

h5_getdataset(filename, dataset)
h5_getdataset(filename, dataset, h5group::String)
h5_getdataset(filename, dataset, batchnr::Int)
h5_getdataset(filename, dataset, batches::AbstractVector)

Returns dataset dataset from all batches, the HDF5-group h5group, batchnr or range of batches.

TraceParticles.h5_getenergies — Function

h5_getenergies(filename, args...; units="eV")

Returns the initial and final energies of a set of batches in a HDF5-file. If batches is not specified, the data from all batches are returned.

TraceParticles.h5_getinitialstate — Function

h5_getinitialstate(filename, args...)

Returns x0, y0, z0, vparal0, and magneticmoment from a test particle ensemble stored in a HDF5-file.

TraceParticles.h5_getfinalstate — Function

h5_getfinalstate(filename, args...)

Returns xf, yf, zf, vparalf, and magneticmoment from a test particle ensemble stored in a HDF5-file.

TraceParticles.save_gcastates — Function

save_gcastates(filename, fields_itp)

Calculates the GCAStates of the initial and final states of an ensamble of test particles.

TraceParticles.create_bifrost_itps — Function

create_bifrost_interpolators(
    brxp::BifrostExperiment,
    snap::Integer,
    itp_type::Interpolations.InterpolationType,
    itp_bc::Interpolations.BoundaryCondition;
    filename::String,
    units::String,
    variables::Vector{String}
)

Create interpolation objects from the MHD variables in a Bifrost snapshot and save them as JLD2-files.

Keyword arguments

normalise::Vector{Bool}

destagger::Vector{Bool}

xzinterpolation::Bool: If true, wrap 2D interpolators in a struct that allows it to be called with three coordinates, using only the first and third coordinate. That is; itp(x,y,z) = itp(x,z).

Dataprocessing

TraceParticles.GCAState — Type

GCAState(
    state            ::Vector{<:Real},
    charge           ::Real,
    mass             ::Real,
    magneticmoment   ::Real,
    time             ::Real,
    kineticenergy    ::Real,
    perpvelocity     ::Real,
    larmorradius     ::Real,
    gyrofrequency    ::Real,
    cosofpitchangle  ::Real,
    exbdrift         ::Vector{<:Real},
    ∇Bdrift          ::Vector{<:Real},
    curvaturedrift   ::Vector{<:Real},
    polarisationdrift::Vector{<:Real},
    mirroracc        ::Real,
    paralacc         ::Real,
    magneticfield    ::Vector{<:Real},
    electricfield    ::Vector{<:Real},
    magneticfieldlength::Real
    )

A struct to hold the state (guiding centre position and its parallel velocity) of a particle in the guiding centre approximation, all its drifts, the external electromagnetic field, relevant parameters, and auxiliary quantities.

Constructors

GCAState(
    state ::Vector{<:Real},
    q     ::Real,
    m     ::Real,
    μ     ::Real,
    electromagneticfield::Any;
    time=nothing,
    )

Constructs a GCAState object from;

state - The guiding centre state; position and parallel velocity

vparal - The parallel velocity of the guiding centre

q - The charge of the particle

m - The mass of the particle

μ - The magnetic moment of the particle

electromagneticfield - A function or functor that returns the electric and magnetic field vectors at an arbitrary position in space.

Keyword arguments:

time - The time of the state

GCAState( state ::Vector{<:Real}, q ::Real, m ::Real, μ ::Real, bfield::Vector{<:Real}, efield::Vector{<:Real} ; time=nothing )

Constructs a partially filled GCAState object, neglecting all drifts except the ExB-drift.

GCAState(
    solution::DataFrame,
    fields::Vector{<:AbstractInterpolation}
    ;
    kwargs...
    )

From DataFrame test-particle solution and the field interpolation vector used in the simulation, construct two ensembles of GCAStates. One for the initial states of the particles and one for the final states.

TraceParticles.get_observable — Function

Get the observable sym from an ODESolution or a vector of ODESolutions at a single time or over a range of times.

Methods

get_observable(vectorofsolutions, symbol; t, kwargs...)
get_observable(solution, symbol, t; kwargs...)
get_observable(solution, symbol; tspan, times, nsteps, kwargs...)

TraceParticles.initialstate_idxs — Function

initialstate_idxs(fname, idxs)

Find the initial state of the particles with the indices idxs from the test particle ensemble stored as HDF5 with the filename fname.

TraceParticles.initialstate_terminationcode — Function

initialstate_terminationcode(fname, terminationcode)

Find the initial state of the particles with a certain terminationcode from the test particle ensemble stored as HDF5 with the filename fname.

TraceParticles.finalstate_terminationcode — Function

finalstate_terminationcode(fname, terminationcode)

Find the final state of the particles with a certain terminationcode from the test particle ensemble stored as HDF5 with the filename fname.

TraceParticles.initialstate_maxiters — Function

initialstate_maxiters(fname::String, nparticles::Int)

Find the initial state of the nparticles particles with most timesteps from the test particle ensemble stored as HDF5 files with the filename fname.

TraceParticles.particlemaskfunction — Function

particlemaskfunction(
    ;
    x0lim=nothing,
    y0lim=nothing
    z0lim=nothing,
    xflim=nothing,
    yflim=nothing
    zflim=nothing,
    timelim=nothing,
    terminationcode=nothing,

) Create a function that filters out particles based on the specified limits. The function accepts a dictionary which it filters based on the key-limit pairs

:x0 for x0lim

:y0 for y0lim

:z0 for z0lim

:xf for xflim

:yf for yflim

:zf for zflim

:tf for timelim

:terminationcode for terminationcode

It assumes the dictionary has the key :x0 to find out how many particles there are in the dataset.

TraceParticles.particlemask — Function

particlemask(df, outerbounds, ignorezones, patternmatches, premask)

Create a mask for the DataFrame df based on the specified conditions.

outerbounds is a dictionary with the keys being the column names and the values being the limits of the outer bounds of that column.

ignorezones is a vector of dictionaries with the keys being the column names and the values being the limits of the "zone" to ignore particles from.

patternmatches is a dictionary with the keys being the column names and the values being a pattern of the column values to match.

TraceParticles.generate_probdistr — Function

generate_probdistr(
    xvalues,
    yvalues,
    zvalues,
    ;
    kwargs...
    )

Generate a 2D probability distribution from discrete values of f_z(x,y).

The algorithm bins the zvalues on the xy-plane and uses the mean z-value of each bin as the value of the distribution function at that point. Finally it normalizes the distribution function to 1 and constructs a linear interpolation object from the discrete probability distribution.

Returns the interpolation object, and the bin edges in x and y.

TraceParticles.replaceparticles — Function

replace_finalstates(fname_original, fname_replacement, original_idxs)
replace_finalstates(fname_original, fname_replacement, original_idxs;
    replace_keys=names)
replace_finalstates(original_df, replacement_df, original_idxs)

Replace the particles in fname_original with the particles in fname_replacement. The indices of the particles to be replaced in fname_original are given by idxs. The length of idxs must be the same as the number of particles in fname_replacement.

TraceParticles.rerun — Function

rerun(u0, mu0, nparticles, fname; kwargs)

Reruns test particles in the experiment defined by fname. The initial positions and velocities of the particles are arbitrary, defined in u0, and their magnetic moments are defined in mu0. Some parameters of the experiment may be adjusted through keyword arguments. These are

Keyword arguments

reltol

abstol

maxiters

alg: solver algorithm

gcatol: Tolerance for the guiding centre approximation.

Example use

```julia fname = "testparticles.h5" npart = 10 sol = rerun(find_nonthermals(fname, npart)[1:2]..., npart, fname)

rerun(
    u0::Vector{<:Vector{<:Real}},
    mu0::Vector{<:Real},
    charge::Real,
    mass::Real,
    fields::Vector{Any},
    tspans::Vector{<:Tuple{Real, Real}},
    ensemble_prob_kwargs::Dict{Symbol, Any},
    solve_kwargs::Dict{Symbol, Any};
    alg=Rosenbrock23(),
    parallelisation=EnsembleSerial(),
)

Rerun GCA particles with initial condition u0 and magnetic moments mu0, using the parameters charge, mass, fields, and tspans. The arguments ensemble_prob_kwargs and solve_kwargs are dictionaries containing keyword arguments for the EnsembleProblem and DifferentialEquations.solve function.

Physics

TraceParticles.get_fullorbit — Function

get_fullorbit(
    magneticfield::AbstractVector,
    electricfield::AbstractVector,
    R            ::AbstractVector,
    vparal       ::Real,
    μ            ::Real,
    mass         ::Real,
    charge       ::Real,
    phaseangle   ::Real,
    )

Get the position and velocity from the guiding centre position R, parallel velocity vparal, and magnetic moment μ of a charged particle with mass and charge in an electromagnetic field.

Requires an arbitrary phaseangle of the gyromotion because we are going from 5 parameters to 6. The phase angle is with respect to vector perpendicular to the magnetic field and guiding centre position vector.

TraceParticles.get_fullorbit! — Function

get_fullorbit!(u::AbstractVector, args...)

In-place version of get_fullorbit.

TraceParticles.get_guidingcentre — Function

get_guidingcentre(
    pos          ::AbstractVector,
    vel          ::AbstractVector,
    magneticfield::AbstractVector,
    electricfield::AbstractVector,
    charge       ::Real,
    mass         ::Real
    )

Calculates and returns the guiding-centre position, parallel velocity, and magnetic moment of a charged particle in a electromagnetic field.

TraceParticles.get_guidingcentre! — Function

get_guidingcentre!(u::AbstractVector, args...)

In-place version of get_guidingcentre.

TraceParticles.kineticenergy — Function

Return the kinetic energy of a particle.

Methods

kineticenergy(velocity::Any, mass::Any)
kineticenergy(velocity::Vector, mass::Any)

kineticenergy(
    parallel_velocity::Real,
    vperp            ::Real,
    exbdrift         ::AbstractVector,
    ∇Bdrift          ::AbstractVector,
    curvaturedrift   ::AbstractVector,
    polarisationdrift::AbstractVector,
    mass             ::Real,
    )

Return the kinetic energy of charged particle given the particle mass, its drifts, and parallel_velocity and perpendicular (vperp) velocity of its guiding centre.

kineticenergy(
     parallel_velocity::Real,
     vperp            ::Real,
     exbdrift         ::AbstractVector,
     mass             ::Real,
)

Return the kinetic energy of charged particle given the particle mass, its E cross B drift, and parallel_velocity and perpendicular (vperp) velocity of its guiding centre. Drifts other than the E cross B drift are neglected.

kineticenergy(
    state::Vector{<:Real}, # Guiding centre state
    q::Real, # Particle charge
    m::Real, # Particle mass
    μ::Real, # Particle magnetic moment
    electromagneticfield::Any,

) Kinetic energy of a charged particle based on its guiding centre, parallel velocity, magnetic moment, and the electromagnetic field in which it is embedded.

kineticenergy(
    gcastate::GCAState
)

Kinetic energy of a charged particle based on its guiding centre state represented by a GCAState object.

TraceParticles.perpendicular_velocity — Function

Return the perpendicular velocity of a charged particle in a electromagnetic field.

Methods

perpendicular_velocity(velocity, magneticfield, electricfield)
perpendicular_velocity(magnetic_moment, mass, magneticfieldstrength)
perpendicular_velocity(velocity, b_vec, vparal)
perpendicular_velocity(R, t, magnetic_moment, mass, emfields)
perpendicular_velocity(pos, vel, t, emfields)

TraceParticles.larmorradius — Function

Return the Larmor radius of a charged particle.

Methods

larmorradius(mass, vperp, charge, magneticfieldstrength)
larmorradius(vel, mass, charge, electricfield, magneticfield)
# Methods with interpolation objects
larmorradius(R, μ, q, m, emfields) # for guiding centre approximation
larmorradius(position, velocity, mass, charge, emfields) # direct solution

TraceParticles.characteristicfieldlength — Function

Return the characteristic length of a field.

Methods

characteristicfieldlength(fieldstrength, fieldstrengthgradient)
characteristicfieldlength(position, emfields)

TraceParticles.scalesratio — Function

Calculates the ratio between the Larmor radius and the characteristic field length of the magnetic field.

Methods

scalesratio(B, ∇B, vperp, q, m)
# Using interpolation object
scalesratio(R, mass, charge, magneticmoment, emfields)
scalesratio(pos, vel, mass, charge, emfields)

Arguments

R, the guiding centre location.

pos, the 3 component particle position.

vel, the 3 component particle velocity.

∇B, the magnitude of the gradient of the magnetic field

B, the magnetic field strength.

vperp, the particle velocity component perpendicular to the magnetic field.

q, the particle charge.

m, the particle mass.

emfields, a functor that returns a tuple of the electric and magnetic field vectors at a given position, in the form

TraceParticles.magneticmoment — Function

Return the magnetic moment of a chargred particle in a magnetic field.

Methods

magneticmoment(perpendicular_velocity, mass, magneticfieldstrength)
magneticmoment(vel, mass, electricfield, magneticfield)
# interpolation object for direct of the Lorentz equation.
magneticmoment(position, velocity, mass, emfields)

Statistics

TraceParticles.maxwellianvelocitysample — Function

maxwellianvelocitysample(
    rng::AbstractRNG,
    temperature_itp::Any,
    mass::Real,
    args...
    ;
    precision::DataType=Float64,
)

Draw a velocity component sample from a Maxwellian distribution with temperature given by a function call with aribtrary arguments, and by a mass.

TraceParticles.rejectionsample — Function

rejectionsample(
    target  ::Any,
    maxvalue::Real,
    domain  ::Vector{<:Tuple{Real,Real}},
    rng     ::AbstractRNG
    )

Sample a point from the target-distribution using rejection sampling. The target function must be callable with dimension (the number of arguments) equal to the length of domain. The maxvalue determines the upper bound of the uniform "proposal" distribution. Make sure sure that this value is equal to or larger than the maximum possible value of the target distribution. If not, the point will not be a true sample of the target distribution.

The function returns the sampled point and the number of rejections.

The rejection sample algorithm draws independent and identically distributed random variables, at the cost of a potentially large number of rejections. The number of rejectiosn icreases with dimension.

rejectionsample(
    target    ::Function,
    maxvalue  ::Real,
    domain    ::Vector{<:Tuple{Real,Real}},
    rng       ::AbstractRNG,
    numsamples::Integer,
    )

Draws numsamples points from the target-distribution using the rejection algorithm. See rejectionsample for more information.

TraceParticles.binmap — Function

binmap(
    data,
    ;
    weights=ones(length(data)),
    mapfunc::Function = (x,w) -> sum(w),
    xvalues=nothing,
    xedges=nothing,
    nbins= 100,
    logx=false,
    )

Maps a binning of the data to the function mapfunc. The default mapping is the sum of the weights of the data in each bin. The defualt weighing is 1.

If no edges or data-corresponding x-values are given, we bin the data in a linear, uniform range between its minimum and maximum value.

Returns the midpoints of the bins and the mapped values of the bins.

binmap(
    xvalues,
    yvalues,
    ;
    data=ones(length(xvalues)),
    mapfunc::Function = x -> length(x),
    nbinsx=100,
    nbinsy=100,
    logx=false,
    logy=false,
    )

Maps a 2D binning of the data to the function mapfunc. Does not support weighing of the data.

Returns the edges of the bins and their mapped values.

binmap(
    xvalues,
    yvalues,
    zvalues,
    ;
    data=ones(length(xvalues)),
    mapfunc::Function = x -> length(x),
    nbins=100
    logx=false,
    logy=false,
    )

Maps a 2D binning of the data to the function mapfunc. Does not support weighing of the data.

Returns the edges of the bins and their mapped values.

Utilities

TraceParticles.ElectromagneticFieldInterpolator — Type

ElectromagneticFieldInterpolator{T<:AbstractVector}(interpolator::T)

A convenience struct for a vector of interpolators representing the components of the electromagnetic field. Following the convention of storing the vector of components that are stored as a JLD2 object [bx, by, bz, ex, ey, ez]. Calling an instance of this struct with spatial coordinates returns the electric and magnetic fields at that point as two SVector{3} vectors.

TraceParticles.StaticInterpolation — Type

StaticInterpolation

An interpolator that calls (x, y, z) if called with (x, y, z, t).

Fields

itp::AbstractInterpolation

TraceParticles.StaticXZInterpolation — Type

StaticXZInterpolation

An interpolator that calls (x, z) if called with (x, y, z, t).

Fields

itp::AbstractInterpolation

TraceParticles.XZInterpolation — Type

XZInterpolation

An interpolator that calls (x, z, t) if called with (x, y, z, t).

Fields

itp::AbstractInterpolation